+++ bench/adapters/micro-promise.js
