+++ bench/fulfilled.js
@@ -5,12 +5,8 @@
 // have actually resolved (e.g. Q promises always resolve in a
 // future turn).  This is a pure, brute force sync code test.
 
-var name = process.argv[2].match(/\/([^\/]+)\.js$/)[1]
-/**
- * Load the adapter
- */
-var lib = require(process.argv[2])
+var createFulfilled = implementation.fulfilled
 
-require('../bench')(name, 10000, function (i) {
-	lib.fulfilled(i)
-})
+module.exports = function(i){
+	createFulfilled(i)
+}
