+++ bench/pending.js
@@ -1,0 +1,3 @@
+// Create a pending promise (10,000 iterations)
+
+module.exports = implementation.createNormal
