--- docs/middle.md
@@ -1,7 +1,0 @@
-```
-
-# Test Results
-
-Each test is sorted from best to worst time. Times are in milliseconds, and Diff is the percentage difference from the best time.
-
-```
