--- docs/tail.md
@@ -1,2 +1,0 @@
-```
-
