--- lib/tests/pending-sequence.js
@@ -1,24 +1,0 @@
-// Sequence from a pending promise
-// Performance of large sequence of then calls from a pending promise
-
-var array = []
-for(var i = 1; i < 10000; i++) {
-    array.push(i)
-}
-
-var name = process.argv[2].match(/\/([^\/]+)\.js$/)[1]
-/**
- * Load the adapter
- */
-var lib = require(process.argv[2])
-
-require('../bench')(name, 1, function () {
-    // Use reduce to chain <iteration> number of promises back
-    // to back.  The final result will only be computed after
-    // all promises have resolved
-    array.reduce(function(promise, nextVal) {
-        return promise.then(function(currentVal) {
-            return currentVal + nextVal
-        })
-    }, lib.pending().promise)
-})
